S6: Local storage
- Browser local storage to store profile data.
- 'react-toastify' for notifications.
