+++ src/index.js
@@ -4,6 +4,7 @@ import { BrowserRouter } from "react-router-dom";
 import { library } from "@fortawesome/fontawesome-svg-core";
 import { fab } from "@fortawesome/free-brands-svg-icons";
 import { fas } from "@fortawesome/free-solid-svg-icons";
+import { ToastContainer } from "react-toastify";
 
 import App from "./App";
 
@@ -16,6 +17,7 @@ const mountTo = document.getElementById("app-root");
 
 ReactDOM.render(
   <BrowserRouter>
+    <ToastContainer />
     <App />
   </BrowserRouter>,
   mountTo
