+++ src/styles.scss
@@ -1,12 +1,13 @@
 @import "bootstrap/scss/bootstrap.scss";
+@import "react-toastify/scss/main.scss";
 
 // Constants
 $color-light: #ffffff;
 $color-light-dull: #f0f0f0;
-$color-pink: #e91e63;
-$color-dark-grey: #2b2b2b;
 $color-light-grey: #b2b2b2;
 $color-dark: #000;
+$color-dark-grey: #2b2b2b;
+$color-pink: #e91e63;
 
 $sizes_map: (
   1: 4rem,
@@ -46,9 +47,9 @@ $sizes_map: (
 }
 
 // Application
-.app-container {
-  height: 100vh;
-  width: 100vw;
+body {
+  height: auto;
+  width: auto;
   background-color: $color-light-dull;
 }
 
@@ -153,8 +154,8 @@ $sizes_map: (
 .forgot-password-page,
 .registration-page,
 .login-page {
-  width: 100vw;
-  height: 100vh;
+  width: 100%;
+  height: 100%;
 
   .user-form {
     width: 95%;
